move tag code from timeseries to taglib
