+++ libs/datasets/tail_filter.py
@@ -6,17 +6,16 @@ import pandas as pd
 from covidactnow.datapublic.common_fields import FieldName
 from covidactnow.datapublic.common_fields import PdFields
 
+from libs.datasets import taglib
 from libs.datasets import timeseries
 
-TagType = timeseries.TagType
-TagField = timeseries.TagField
+TagType = taglib.TagType
+TagField = taglib.TagField
 
 
 @dataclasses.dataclass
 class TailFilter:
-    _annotations: timeseries.TagCollection = dataclasses.field(
-        default_factory=timeseries.TagCollection
-    )
+    _annotations: taglib.TagCollection = dataclasses.field(default_factory=taglib.TagCollection)
 
     # Counts that track what the filter has done.
     skipped_too_short: int = 0
@@ -107,10 +106,10 @@ class TailFilter:
             # series_in.iat[truncate_at] is the first value *not* returned
             assert TailFilter.FILTER_DATES_OLDEST <= truncate_at <= -1
             if count_observation_diff_under_threshold < TailFilter.COUNT_OBSERVATION_LONG:
-                annotation_type = timeseries.CumulativeTailTruncated
+                annotation_type = taglib.CumulativeTailTruncated
                 self.truncated += 1
             else:
-                annotation_type = timeseries.CumulativeLongTailTruncated
+                annotation_type = taglib.CumulativeLongTailTruncated
                 self.long_truncated += 1
             # Currently one annotation is created per series. Maybe it makes more sense to add
             # one for each dropped observation / real value?
