+++ tests/libs/datasets/taglib_test.py
@@ -3,6 +3,7 @@ import inspect
 from typing import Iterable
 from typing import Type
 
+from libs.datasets import taglib
 from libs.datasets import timeseries
 
 
@@ -24,17 +25,17 @@ def _get_subclass_tag_types(cls):
 
 def test_all_tag_subclasses_accounted_for():
     """Checks that every subclass of TagInTimeseries has exactly one TagType enum value."""
-    subclass_tag_types = _get_subclass_tag_types(timeseries.TagInTimeseries)
-    assert set(timeseries.TagType) == set(subclass_tag_types)
-    assert len(list(timeseries.TagType)) == len(subclass_tag_types)
+    subclass_tag_types = _get_subclass_tag_types(taglib.TagInTimeseries)
+    assert set(taglib.TagType) == set(subclass_tag_types)
+    assert len(list(taglib.TagType)) == len(subclass_tag_types)
 
 
 def test_annotation_tag_types():
-    annotation_tag_types = _get_subclass_tag_types(timeseries.AnnotationWithDate)
+    annotation_tag_types = _get_subclass_tag_types(taglib.AnnotationWithDate)
     assert sorted(annotation_tag_types) == sorted(timeseries.ANNOTATION_TAG_TYPES)
 
 
 def test_tag_type_to_class():
-    assert set(timeseries.TAG_TYPE_TO_CLASS.keys()) == set(timeseries.TagType)
-    for tag_type, tag_type_class in timeseries.TAG_TYPE_TO_CLASS.items():
+    assert set(taglib.TAG_TYPE_TO_CLASS.keys()) == set(taglib.TagType)
+    for tag_type, tag_type_class in taglib.TAG_TYPE_TO_CLASS.items():
         assert tag_type_class.TAG_TYPE is tag_type
