+++ tests/libs/datasets/tail_filter_test.py
@@ -2,7 +2,7 @@ import pytest
 
 from covidactnow.datapublic.common_fields import CommonFields
 
-from libs.datasets.timeseries import TagType
+from libs.datasets.taglib import TagType
 from libs.datasets.tail_filter import TailFilter
 
 from tests import test_helpers
